feat: initialize Vite setup with TypeScript and React
- Added vite-env.d.ts for Vite client types.
- Created tsconfig.json with TypeScript compiler options.
- Added vite.config.ts for Vite configuration with React plugin.
- Removed obsolete .gitignore and CSS files related to layout and share pages.
- Deleted Header.tsx and SharePage.tsx components along with their styles.
