+++ .github/workflows/deploy-site-to-pages.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches: [main, master]
     paths:
-      - 'site/**'
+      - 'packages/site/**'
   workflow_dispatch:
 
 # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
@@ -37,17 +37,17 @@ jobs:
         uses: actions/configure-pages@v5
 
       - name: Install dependencies
-        run: cd site && bun install
+        run: cd packages/site && bun install
 
       - name: Build site
-        run: cd site && bun run build
+        run: cd packages/site && bun run build
         env:
           BASE_URL: /ai-clash/
 
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
-          path: site/dist
+          path: packages/site/dist
 
   deploy:
     environment:
@@ -59,4 +59,4 @@ jobs:
     steps:
       - name: Deploy to GitHub Pages
         id: deployment
-        uses: actions/deploy-pages@v4
+        uses: actions/deploy-pages@v5
