+++ packages/site/package.json
@@ -9,21 +9,21 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@ant-design/icons": "^6.2.3",
+    "@ant-design/icons": "^6.2.5",
     "@ant-design/x": "^2.7.0",
     "@ant-design/x-markdown": "^2.7.0",
-    "@lobehub/icons": "^5.8.0",
+    "@lobehub/icons": "^5.10.0",
     "@vitejs/plugin-react-swc": "^4.3.1",
-    "antd": "^6.4.2",
+    "antd": "^6.4.3",
     "antd-style": "^4.1.0",
     "typescript": "~5.9.3",
-    "vite": "^7.3.3",
-    "react": "^19.2.6",
-    "react-dom": "^19.2.6",
+    "vite": "^7.3.5",
+    "react": "^19.2.7",
+    "react-dom": "^19.2.7",
     "react-markdown": "^10.1.0"
   },
   "devDependencies": {
-    "@types/react": "^19.2.14",
+    "@types/react": "^19.2.16",
     "@types/react-dom": "^19.2.3"
   }
 }
