+++ packages/site/scripts/copy-routes.mjs
