+++ packages/site/src/App.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { App as AntApp, ConfigProvider, Layout } from 'antd'
 import enUS from 'antd/locale/en_US'
 import zhCN from 'antd/locale/zh_CN'
@@ -8,11 +8,18 @@ import { createAntTheme } from './app/theme'
 import type { ThemeMode } from './app/theme'
 import { Footer } from './layout/Footer'
 import { Header } from './layout/Header'
-import { ChangelogPage, DownloadPage, HomePage, PrivacyPage, SharePage } from './pages'
+import { ChangelogPage, DownloadPage, HomePage, PrivacyPage } from './pages'
 
 const { Content } = Layout
 
-function Page({ locale, page }: { locale: Locale; page: string }) {
+function Page({
+  locale,
+  page
+}: {
+  locale: Locale
+  page: string
+  onLoginRequired: () => void
+}) {
   switch (page) {
     case '/download':
       return <DownloadPage locale={locale} />
@@ -26,39 +33,47 @@ function Page({ locale, page }: { locale: Locale; page: string }) {
 }
 
 export function App() {
-  const path = useMemo(() => stripBasePath(window.location.pathname), [])
+  const [path, setPath] = useState(() => stripBasePath(window.location.pathname))
   const [themeMode, setThemeMode] = useState<ThemeMode>(() => {
     const saved = window.localStorage.getItem('ai-clash-site-theme')
     if (saved === 'dark' || saved === 'light') return saved
     return window.matchMedia?.('(prefers-color-scheme: dark)').matches ? 'dark' : 'light'
   })
+  const locale = getInitialLocale(path)
+  const page = stripLocale(path)
+  const isSharePage = page === '/share' || page.startsWith('/share/')
+  const isAppPage = page === '/chat' || isSharePage
+
+  useEffect(() => {
+    const updatePath = () => setPath(stripBasePath(window.location.pathname))
+    window.addEventListener('popstate', updatePath)
+    return () => window.removeEventListener('popstate', updatePath)
+  }, [])
 
   useEffect(() => {
     document.documentElement.dataset.theme = themeMode
     window.localStorage.setItem('ai-clash-site-theme', themeMode)
   }, [themeMode])
 
-  const locale = getInitialLocale(path)
-  const page = stripLocale(path)
   const antLocale = locale === 'en' ? enUS : zhCN
   const antTheme = createAntTheme(themeMode)
-  const pageContent =
-    page === '/share' || page.startsWith('/share/')
-      ? <SharePage locale={locale} themeMode={themeMode} shareId={page.startsWith('/share/') ? page.slice('/share/'.length) : undefined} />
-      : <Page locale={locale} page={page} />
+  const pageContent = <Page locale={locale} page={page} onLoginRequired={() => false} />
 
   return (
     <ConfigProvider locale={antLocale} theme={antTheme}>
       <Layout className="site-layout">
         <AntApp>
-          <Header
-            locale={locale}
-            path={path}
-            themeMode={themeMode}
-            onToggleTheme={() => setThemeMode(themeMode === 'dark' ? 'light' : 'dark')}
-          />
+          {!isAppPage ? (
+            <Header
+              locale={locale}
+              path={path}
+              themeMode={themeMode}
+              onToggleTheme={() => setThemeMode(themeMode === 'dark' ? 'light' : 'dark')}
+              onLoginRequired={() => false}
+            />
+          ) : null}
           <Content>{pageContent}</Content>
-          <Footer />
+          {isAppPage ? null : <Footer />}
         </AntApp>
       </Layout>
     </ConfigProvider>
