+++ packages/site/src/app/api.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL = (import.meta.env.VITE_API_BASE_URL || 'https://ai-clash-service.snewbie.site').replace(/\/+$/, '')
