+++ packages/site/src/pages/index.ts
@@ -2,4 +2,3 @@ export { ChangelogPage } from './ChangelogPage'
 export { DownloadPage } from './DownloadPage'
 export { HomePage } from './HomePage'
 export { PrivacyPage } from './PrivacyPage'
-export { SharePage } from './SharePage'
