+++ packages/site/src/styles.css
@@ -2,4 +2,3 @@
 @import './styles/layout.css';
 @import './styles/home.css';
 @import './styles/docs.css';
-@import './styles/share.css';
