+++ packages/site/src/styles/home.css
@@ -49,8 +49,10 @@
   display: inline-flex;
   align-items: center;
   justify-content: center;
+  height: 48px !important;
   min-height: 48px;
-  padding: 0 20px;
+  padding: 0 20px !important;
+  padding-block: 0 !important;
   border: 1px solid transparent;
   border-radius: 24px;
   font-size: 16px;
@@ -280,7 +282,9 @@
 
   .button.ant-btn {
     width: 100%;
+    height: 46px !important;
     min-height: 46px;
+    padding-block: 0 !important;
     font-size: 15px;
   }
 }
