--- site/src/styles/share.css
@@ -1,252 +1,0 @@
-.share-page {
-  display: flex;
-  min-height: auto;
-  background: var(--vp-c-bg-soft);
-}
-
-.share-panel {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  max-width: none;
-  min-height: auto;
-  overflow: visible;
-  background: var(--vp-c-bg);
-  color: var(--vp-c-text-1);
-}
-
-.share-panel .x-markdown,
-.share-panel .x-markdown-light,
-.share-panel .x-markdown-dark {
-  font-size: 14px;
-  line-height: 1.65;
-}
-
-.share-panel .ant-bubble-content-filled {
-  background: var(--vp-c-bg-soft);
-}
-
-.share-panel .x-markdown ul,
-.share-panel .x-markdown ol,
-.share-panel .x-markdown p,
-.share-panel .x-markdown li,
-.share-panel .x-markdown-light ul,
-.share-panel .x-markdown-light ol,
-.share-panel .x-markdown-light p,
-.share-panel .x-markdown-light li,
-.share-panel .x-markdown-dark ul,
-.share-panel .x-markdown-dark ol,
-.share-panel .x-markdown-dark p,
-.share-panel .x-markdown-dark li {
-  margin: 0;
-}
-
-.share-panel .x-markdown h1,
-.share-panel .x-markdown h2,
-.share-panel .x-markdown h3,
-.share-panel .x-markdown h4,
-.share-panel .x-markdown h5,
-.share-panel .x-markdown h6,
-.share-panel .x-markdown-light h1,
-.share-panel .x-markdown-light h2,
-.share-panel .x-markdown-light h3,
-.share-panel .x-markdown-light h4,
-.share-panel .x-markdown-light h5,
-.share-panel .x-markdown-light h6,
-.share-panel .x-markdown-dark h1,
-.share-panel .x-markdown-dark h2,
-.share-panel .x-markdown-dark h3,
-.share-panel .x-markdown-dark h4,
-.share-panel .x-markdown-dark h5,
-.share-panel .x-markdown-dark h6 {
-  margin-top: 16px;
-}
-
-.share-floating-btn-text {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-  height: 32px;
-  padding: 0 12px;
-  border: 1px solid var(--vp-c-divider);
-  border-radius: 16px;
-  background: var(--vp-c-header);
-  backdrop-filter: blur(12px);
-  color: var(--vp-c-text-2);
-  cursor: pointer;
-  font-size: 13px;
-  font-weight: 600;
-  transition: all 0.2s ease;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.06);
-}
-
-.share-chat-list {
-  display: flex;
-  flex: 0 1 auto;
-  min-height: 0;
-  flex-direction: column;
-  overflow: visible;
-  padding-block: 24px;
-}
-
-.share-state-view {
-  display: flex;
-  min-height: 52vh;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  padding: 72px 24px;
-  text-align: center;
-}
-
-.share-state-view__mark {
-  display: inline-flex;
-  width: 44px;
-  height: 44px;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid var(--vp-c-divider);
-  border-radius: 999px;
-  background: var(--vp-c-bg-soft);
-  color: var(--vp-c-text-2);
-  font-size: 18px;
-}
-
-.share-state-view__mark.is-loading {
-  color: var(--vp-c-brand-1);
-}
-
-.share-state-view h1 {
-  margin: 18px 0 8px;
-  color: var(--vp-c-text-1);
-  font-size: 24px;
-  font-weight: 700;
-  letter-spacing: 0;
-}
-
-.share-state-view p {
-  max-width: 420px;
-  margin: 0;
-  color: var(--vp-c-text-2);
-  font-size: 14px;
-  line-height: 1.7;
-}
-
-.share-state-view__button {
-  display: inline-flex;
-  height: 36px;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
-  margin-top: 24px;
-  padding: 0 14px;
-  border: 1px solid var(--vp-c-divider);
-  border-radius: 999px;
-  background: var(--vp-c-bg);
-  color: var(--vp-c-text-1);
-  font-size: 13px;
-  font-weight: 600;
-  text-decoration: none;
-  box-shadow: 0 1px 4px rgba(0, 0, 0, 0.06);
-}
-
-.share-chat-list .ant-bubble-list {
-  flex: 0 1 auto;
-  overflow-y: visible;
-  padding-bottom: 8px;
-}
-
-.share-chat-list .ant-bubble + .ant-bubble {
-  margin-top: 12px;
-}
-
-.share-provider-header {
-  display: flex;
-  width: auto;
-  align-items: center;
-  gap: 6px;
-  padding: 8px 12px;
-  margin: -8px -12px 8px;
-  border: 0;
-  border-radius: 8px 8px 0 0;
-  background: transparent;
-  color: inherit;
-  cursor: pointer;
-  font: inherit;
-  text-align: left;
-  user-select: none;
-  transition: background 0.15s ease;
-}
-
-.share-provider-header:hover {
-  background: var(--vp-c-bg-soft);
-}
-
-.share-provider-header__arrow {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  color: var(--vp-c-text-3);
-  font-size: 12px;
-  transform: rotate(90deg);
-  transition: transform 0.2s ease;
-}
-
-.share-provider-header__arrow.is-collapsed {
-  transform: rotate(0deg);
-}
-
-.share-provider-header__content {
-  display: flex;
-  align-items: center;
-  flex: 1;
-  min-width: 0;
-  gap: 6px;
-}
-
-.share-provider-header__icon {
-  flex-shrink: 0;
-  font-size: 14px;
-}
-
-.share-provider-header__name {
-  color: var(--vp-c-text-1);
-  font-size: 13px;
-  font-weight: 600;
-  opacity: 0.9;
-  white-space: nowrap;
-}
-
-.share-provider-header__status {
-  overflow: hidden;
-  color: var(--vp-c-text-3);
-  font-size: 11px;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.share-bubble-content-hidden > .ant-bubble-body {
-  padding-top: 0 !important;
-  padding-bottom: 0 !important;
-}
-
-.share-bubble-content-hidden .ant-bubble-content {
-  display: none !important;
-}
-
-@media (max-width: 520px) {
-  .share-page {
-    background: var(--vp-c-bg);
-  }
-
-  .share-panel {
-    width: 100vw;
-  }
-
-  .share-provider-header__name {
-    display: none;
-  }
-}
